+++ src/System/MQ/Encoding/MessagePack.hs
@@ -11,7 +11,8 @@ import           Control.Monad.Except (throwError)
 import qualified Data.ByteString      as BS (ByteString)
 import qualified Data.ByteString.Lazy as BSL (fromStrict, toStrict)
 import qualified Data.MessagePack     as MP (MessagePack (..), pack, unpack)
-import           System.MQ.Monad      (MQError (..), MQMonad)
+import           System.MQ.Monad      (MQMonad)
+import           System.MQ.Error      (MQError (..), errorEncoding)
 import           Text.Printf          (printf)
 
 -- | Packs something from 'MP.MessagePack' to 'BS.ByteString'.
@@ -31,4 +32,4 @@ unpack = MP.unpack . BSL.fromStrict
 unpackM :: MP.MessagePack a => BS.ByteString -> MQMonad a
 unpackM bs@(unpack -> m) = maybe (throwError err) pure m
   where
-    err = MQProtocolError . printf "could not unpack MessagePack: %s" . show $ bs
+    err = MQError errorEncoding . printf "could not unpack MessagePack: %s" . show $ bs