+++ src/System/MQ/Error.hs
@@ -1,0 +1,7 @@
+module System.MQ.Error
+  (
+    module System.MQ.Error.Internal.Types
+  ) where
+
+import System.MQ.Error.Internal.Types
+import System.MQ.Error.Internal.Instances ()