+++ src/System/MQ/Protocol.hs
@@ -5,11 +5,9 @@ module System.MQ.Protocol
   , module System.MQ.Protocol.Internal.Functions
   , module System.MQ.Protocol.Internal.Tag
   , module System.MQ.Protocol.Class
-  , module System.MQ.Protocol.Error
   ) where
 
 import           System.MQ.Protocol.Class
-import           System.MQ.Protocol.Error
 import           System.MQ.Protocol.Internal.Condition
 import           System.MQ.Protocol.Internal.Functions
 import           System.MQ.Protocol.Internal.Instances ()